tests/rgw: refactor git clone
